docs: document thread sender interface
